src/requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
astroid==2.9.3
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.74.1
fastapi-pagination==0.9.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
loguru==0.6.0
Mako==1.2.0
MarkupSafe==2.1.0
mccabe==0.6.1
passlib==1.7.4
platformdirs==2.5.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pylint==2.12.2
PyMySQL==1.0.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
uvicorn==0.17.5
win32-setctime==1.1.0
wrapt==1.13.3
